requirements.txt
# === Core Framework ===
Flask==3.0.0
Flask-Session==0.5.0
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
gunicorn==22.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# === Database ===
psycopg[binary]>=3.1.0

# === Auth ===
gotrue>=2.0.0
PyJWT>=2.8.0

# === Task Queue ===
celery[redis]==5.4.0
redis==5.0.0

# === Web Scraping / HTTP ===
requests==2.31.0
beautifulsoup4==4.12.2
lxml
urllib3==2.0.7

# === AI / LLM ===
groq>=0.12.0

# === ML ===
scikit-learn>=1.3.2
pandas>=2.1.4
joblib>=1.3.2

# === Reports ===
fpdf2==2.7.6

# === Monitoring ===
prometheus-client>=0.19.0

# === Testing ===
pytest==7.4.3
pytest-cov==4.1.0
